UBUNTU: SAUCE: (drop after 2.6.31) dell-laptop: Trigger rfkill updates on wifi toggle switch press
OriginalLocation: http://patchwork.kernel.org/patch/38439/
BugLink: http://bugs.launchpad.net/bugs/430809

rtg - This patch is on its way upstream for 2.6.33, but will likely land in a different form
after the keyboard filter is implemented in i8042.

Dell hardware sends the rfkill hardware killswitch event via the keyboard
controller, even if it's a physical toggle switch on the side of the
machine. Add support for catching the input event and updating the rfkill
state, allowing userspace to receive notifications that the change has
occurred.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>