UBUNTU: SAUCE: (drop after 2.6.31) input: Add support for filtering input events
OriginalLocation: http://patchwork.kernel.org/patch/38440/
BugLink: http://bugs.launchpad.net/bugs/430809

rtg - Matthew says that this patch will get moved to the i8042 driver
for 2.6.33. However, given its limited use in this kernel (dell-laptop only so far),
it should be benign for those platform drivers that do not register
a keyboard filter.

Devices occasionally use the keyboard controller to send hardware
notifications that should be handled by the kernel rather than userspace.
These can be handled in kernel by registering an input handler, but the
event will still bubble up to userspace where it may cause confusion. This
patch adds support for the kernel to register filters, allowing it to
indicate that the event should be consumed by the hardware-specific driver
rather than passed to other input handlers. The assumption is that, as this
is hardware specific, there should be no need to have more than one filter -
similarly, if an input device is grabbed, we assume that people know what
they're doing and don't pass it to the filter.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>