UBUNTU: [Upstream] elevator: fix fastfail checks to allow merge of readahead requests
BugLink: http://bugs.launchpad.net/bugs/444915

When a readahead request is injected into the elevator we construct a
bio as below, treating it as fastfail:

    void init_request_from_bio(struct request *req, struct bio *bio)
    [...]
        if (bio_rw_ahead(bio))
                req->cmd_flags |= (REQ_FAILFAST_DEV | REQ_FAILFAST_TRANSPORT |
                                   REQ_FAILFAST_DRIVER);
    [...]

However when merging requests in the elevator we only allow merges of
new bios with matching fastfail attributes:

    int elv_rq_merge_ok(struct request *rq, struct bio *bio)
    [...]
        if (!bio_failfast_dev(bio)       != !blk_failfast_dev(rq) ||
            !bio_failfast_transport(bio) != !blk_failfast_transport(rq) ||
            !bio_failfast_driver(bio)    != !blk_failfast_driver(rq))
    [...]

This check occurs against the original bio fastfail bits ignoring the
override and thus preventing merge of an existing readahead request with
one in the queue.

Modify the merge check to take account of the fact that all readahead
bios will be treated as fastfail.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>